restructure datatypes.py into separate directory of generator/handler scripts
