+++ src/dx/utils/__init__.py
@@ -1,5 +1,2 @@
-from .datatypes import *
-from .date_time import *
 from .formatting import *
-from .geometry import *
 from .tracking import *
